move createOrEditCategoryViewModel out of the fragment
