+++ app/src/main/java/com/infomaniak/drive/ui/fileList/fileDetails/CreateOrEditCategoryFragment.kt
@@ -27,28 +27,17 @@ import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.liveData
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.GridLayoutManager
 import com.infomaniak.drive.MatomoDrive.trackCategoriesEvent
 import com.infomaniak.drive.R
-import com.infomaniak.drive.data.api.ApiRepository
 import com.infomaniak.drive.data.api.ErrorCode.Companion.translateError
-import com.infomaniak.drive.data.cache.DriveInfosController
-import com.infomaniak.drive.data.cache.FileController
-import com.infomaniak.drive.data.models.File
-import com.infomaniak.drive.data.models.drive.Category
 import com.infomaniak.drive.ui.MainViewModel
 import com.infomaniak.drive.ui.fileList.fileDetails.CreateOrEditCategoryAdapter.Companion.COLORS
-import com.infomaniak.drive.utils.find
 import com.infomaniak.drive.utils.getScreenSizeInDp
-import com.infomaniak.lib.core.models.ApiResponse
 import com.infomaniak.lib.core.utils.*
 import kotlinx.android.synthetic.main.fragment_create_or_edit_category.*
-import kotlinx.coroutines.Dispatchers
 import kotlin.math.max
 
 class CreateOrEditCategoryFragment : Fragment() {
@@ -58,10 +47,6 @@ class CreateOrEditCategoryFragment : Fragment() {
     private val navigationArgs: CreateOrEditCategoryFragmentArgs by navArgs()
 
     private val colorsAdapter: CreateOrEditCategoryAdapter by lazy { CreateOrEditCategoryAdapter() }
-    private val driveId: Int
-        get() = files.first().driveId
-
-    private lateinit var files: List<File>
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View =
         inflater.inflate(R.layout.fragment_create_or_edit_category, container, false)
@@ -69,11 +54,6 @@ class CreateOrEditCategoryFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) = with(navigationArgs) {
         super.onViewCreated(view, savedInstanceState)
 
-        files = filesId?.map { fileId -> FileController.getFileById(fileId) }?.filterNotNull() ?: listOf()
-        if (files.isEmpty()) {
-            findNavController().popBackStack()
-            return@with
-        }
         setCategoryName()
         configureColorsAdapter()
         toolbar.setNavigationOnClickListener { findNavController().popBackStack() }
@@ -82,6 +62,10 @@ class CreateOrEditCategoryFragment : Fragment() {
         appBarTitle.title = getString(if (isCreateCategory) R.string.createCategoryTitle else R.string.editCategoryTitle)
         editCategoryWarning.isGone = isCreateCategory
         setSaveButton(isCreateCategory)
+
+        createOrEditCategoryViewModel.init(filesId).observe(viewLifecycleOwner) { hasNoFiles ->
+            if (hasNoFiles) findNavController().popBackStack()
+        }
     }
 
     override fun onConfigurationChanged(newConfig: Configuration) {
@@ -133,11 +117,12 @@ class CreateOrEditCategoryFragment : Fragment() {
         val screenWidth = requireActivity().getScreenSizeInDp().x
         val margins = resources.getDimensionPixelSize(R.dimen.marginStandardSmall).toDp() * 2
         val expectedItemSize = resources.getDimensionPixelSize(R.dimen.coloredChipSize).toDp()
+
         return max(minNumberOfColumns, (screenWidth - margins) / expectedItemSize)
     }
 
-    private fun createCategory() {
-        createOrEditCategoryViewModel.createCategory(
+    private fun createCategory() = with(createOrEditCategoryViewModel) {
+        createCategory(
             driveId = driveId,
             name = categoryNameValueInput.text.toString(),
             color = COLORS[colorsAdapter.selectedPosition],
@@ -153,8 +138,8 @@ class CreateOrEditCategoryFragment : Fragment() {
         }
     }
 
-    private fun addCategory(categoryId: Int) {
-        mainViewModel.manageCategory(driveId, categoryId, files, true).observe(viewLifecycleOwner) { apiResponse ->
+    private fun addCategory(categoryId: Int) = with(createOrEditCategoryViewModel) {
+        mainViewModel.manageCategory(driveId, categoryId, selectedFiles, true).observe(viewLifecycleOwner) { apiResponse ->
             with(apiResponse) {
                 if (isSuccess()) {
                     findNavController().popBackStack()
@@ -166,8 +151,8 @@ class CreateOrEditCategoryFragment : Fragment() {
         }
     }
 
-    private fun editCategory(categoryId: Int) {
-        createOrEditCategoryViewModel.editCategory(
+    private fun editCategory(categoryId: Int) = with(createOrEditCategoryViewModel) {
+        editCategory(
             driveId = driveId,
             categoryId = categoryId,
             name = categoryNameValueInput.text.toString().let { it.ifEmpty { null } },
@@ -179,34 +164,6 @@ class CreateOrEditCategoryFragment : Fragment() {
                 } else {
                     saveButton.hideProgress(R.string.buttonSave)
                     SnackbarUtils.showSnackbar(requireView(), translateError())
-                }
-            }
-        }
-    }
-
-    internal class CreateOrEditCategoryViewModel : ViewModel() {
-
-        fun createCategory(driveId: Int, name: String, color: String): LiveData<ApiResponse<Category>> {
-            return liveData(Dispatchers.IO) {
-                with(ApiRepository.createCategory(driveId, name, color)) {
-                    if (isSuccess()) DriveInfosController.updateDrive { it.categories.add(data) }
-                    emit(this)
-                }
-            }
-        }
-
-        fun editCategory(driveId: Int, categoryId: Int, name: String?, color: String): LiveData<ApiResponse<Category>> {
-            return liveData(Dispatchers.IO) {
-                with(ApiRepository.editCategory(driveId, categoryId, name, color)) {
-                    if (isSuccess()) {
-                        DriveInfosController.updateDrive { localDrive ->
-                            localDrive.categories.apply {
-                                find(categoryId)?.deleteFromRealm()
-                                add(data)
-                            }
-                        }
-                    }
-                    emit(this)
                 }
             }
         }
